poc-rag/requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
transformers
langchain-community
streamlit
python-dotenv
faiss-cpu
